Merge branch 'master' of ssh://bitbucket.org/bitbucket_vpl/vplanet

+++ eqtide.c
@@ -70,7 +70,7 @@ void BodyCopyEqtide(BODY *dest,BODY *src,int iTideModel,int iBody) {
 
 void InitializeBodyEqtide(BODY *body,CONTROL *control,UPDATE *update,int iBody,int iModule) {
   body[iBody].iaTidePerts = malloc(body[iBody].iTidePerts*sizeof(int));
-  body[iBody].daDoblDtEqtide = malloc(control->Evolve.iNumBodies*sizeof(int));
+  body[iBody].daDoblDtEqtide = malloc(control->Evolve.iNumBodies*sizeof(double));
 }
 
 void InitializeUpdateTmpBodyEqtide(BODY *body,CONTROL *control,UPDATE *update,int iBody) {
@@ -80,7 +80,7 @@ void InitializeUpdateTmpBodyEqtide(BODY *body,CONTROL *control,UPDATE *update,in
   control->Evolve.tmpBody[iBody].dTidalZ = malloc(control->Evolve.iNumBodies*sizeof(double));
   
   control->Evolve.tmpBody[iBody].iaTidePerts = malloc(body[iBody].iTidePerts*sizeof(int));
-  control->Evolve.tmpBody[iBody].daDoblDtEqtide = malloc(control->Evolve.iNumBodies*sizeof(int));
+  control->Evolve.tmpBody[iBody].daDoblDtEqtide = malloc(control->Evolve.iNumBodies*sizeof(double));
 
   if (control->Evolve.iEqtideModel == CPL) {
     control->Evolve.tmpBody[iBody].iTidalEpsilon = malloc(control->Evolve.iNumBodies*sizeof(int*));
